+++ packages/fuselage/src/components/Contextualbar/ContextualbarHeader.tsx
@@ -33,5 +33,4 @@ const ContextualbarHeader = ({
     </Box>
   </Box>
 );
-
 export default memo(ContextualbarHeader);