+++ packages/fuselage/src/components/ContextualbarV2/ContextualbarAction.tsx
@@ -1,0 +1,18 @@
+import type { ReactElement, ComponentProps } from 'react';
+import React, { memo } from 'react';
+
+import type { Icon } from '..';
+import { IconButton } from '..';
+
+type ContextualbarActionProps = {
+  name: ComponentProps<typeof Icon>['name'];
+} & Omit<ComponentProps<typeof IconButton>, 'icon'>;
+
+const ContextualbarAction = ({
+  name,
+  ...props
+}: ContextualbarActionProps): ReactElement => (
+  <IconButton {...props} small flexShrink={0} icon={name} />
+);
+
+export default memo(ContextualbarAction);