+++ packages/fuselage/src/components/ContextualbarV2/index.ts
@@ -1,0 +1,27 @@
+import Contextualbar from './Contextualbar';
+import ContextualbarAction from './ContextualbarAction';
+import ContextualbarActions from './ContextualbarActions';
+import ContextualbarButton from './ContextualbarButton';
+import ContextualbarContent from './ContextualbarContent';
+import ContextualbarEmptyContent from './ContextualbarEmptyContent';
+import ContextualbarFooter from './ContextualbarFooter';
+import ContextualbarHeader from './ContextualbarHeader';
+import ContextualbarIcon from './ContextualbarIcon';
+import ContextualbarSection from './ContextualbarSection';
+import ContextualbarSkeleton from './ContextualbarSkeleton';
+import ContextualbarTitle from './ContextualbarTitle';
+
+export {
+  Contextualbar as ContextualbarV2,
+  ContextualbarAction as ContextualbarV2Action,
+  ContextualbarActions as ContextualbarV2Actions,
+  ContextualbarButton as ContextualbarV2Button,
+  ContextualbarContent as ContextualbarV2Content,
+  ContextualbarEmptyContent as ContextualbarV2EmptyContent,
+  ContextualbarFooter as ContextualbarV2Footer,
+  ContextualbarHeader as ContextualbarV2Header,
+  ContextualbarIcon as ContextualbarV2Icon,
+  ContextualbarSkeleton as ContextualbarV2Skeleton,
+  ContextualbarTitle as ContextualbarV2Title,
+  ContextualbarSection as ContextualbarV2Section,
+};