+++ packages/fuselage/src/components/Modal/ModalClose.tsx
@@ -7,5 +7,5 @@ import { IconButton } from '../Button';
 export type ModalCloseProps = ComponentProps<typeof Box>;
 
 export const ModalClose = (props: ModalCloseProps) => (
-  <IconButton small {...props} icon='cross' />
+  <IconButton aria-label='Close' {...props} small icon='cross' />
 );