+++ packages/fuselage/src/components/Table/TableSelection/TableSelectionButton.tsx
@@ -1,0 +1,10 @@
+import type { ComponentProps } from 'react';
+import React from 'react';
+
+import { Button } from '../../Button';
+
+type TableSelectionButtonProps = ComponentProps<typeof Button>;
+
+export const TableSelectionButton = (props: TableSelectionButtonProps) => (
+  <Button small flexShrink={0} {...props} />
+);