+++ packages/fuselage/src/components/Table/TableSelection/TableSelectionButtonGroup.tsx
@@ -1,0 +1,8 @@
+import type { ComponentProps } from 'react';
+import React from 'react';
+
+import { ButtonGroup } from '../../ButtonGroup';
+
+export const TableSelectionButtonGroup = (
+  props: ComponentProps<typeof ButtonGroup>
+) => <ButtonGroup {...props} />;