+++ packages/fuselage/src/components/index.ts
@@ -18,6 +18,7 @@ export * from './CheckBox';
 export * from './Chevron';
 export { default as CodeSnippet } from './CodeSnippet';
 export * from './Contextualbar';
+export * from './ContextualbarV2';
 export { default as Chip } from './Chip';
 export * from './Divider';
 export * from './Dropdown';