+++ packages/fuselage-toastbar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocket.chat/fuselage-toastbar
 
+## 0.32.0
+
+### Minor Changes
+
+- [#1389](https://github.com/RocketChat/fuselage/pull/1389) [`8c80efc47`](https://github.com/RocketChat/fuselage/commit/8c80efc47f3779869875336da88b3185c06a508b) Thanks [@dougfabris](https://github.com/dougfabris)! - feat(fuselage-toastbar): Add RTL support
+
 ## 0.31.26
 
 ### Patch Changes